Fix earth time going backward. Revert to supporting Java 11
r5261
2020-01-22

## DISCUSSION
1. https://github.com/mars-sim/mars-sim/issues/267

## CHANGE
1. Revert to supporting Java 11 due to Java 12 denying
   weblaf's reflective access.
2. Modify README.md to show the support of Java 11 only.
3. Unify .gitignore content across all sub-maven modules.

## FIX
1. Prevent earth time from going backward
   - Use int for currentTR
   - Use long variable for earthMillis

+++ mars-sim-console/src/main/java/org/mars/sim/console/TimeRatioMenu.java
@@ -58,7 +58,7 @@ public class TimeRatioMenu implements BiConsumer<TextIO, RunnerData> {
         
 		if (speedInt >= 0 && speedInt <= 14) {
         	double ratio = Math.pow(2, speedInt);
-        	masterClock.setTimeRatio(ratio);   
+        	masterClock.setTimeRatio((int)ratio);   
         	
             terminal.printf(System.lineSeparator() 
             		+ "The new simulation speed becomes %d"//  -->  New Time-Ratio = 2^speed = %dx"
